+++ src/main/kotlin/com/atlassian/performance/tools/aws/DefaultAmiNameRegistry.kt
@@ -1,5 +1,0 @@
-package com.atlassian.performance.tools.aws
-
-internal object DefaultAmiNameRegistry {
-    const val hvmSsdUbuntuFocal20_04Amd64Server20220610 = "ubuntu/images/hvm-ssd/ubuntu-focal-20.04-amd64-server-20220610"
-}