+++ src/main/kotlin/com/atlassian/performance/tools/aws/api/ami/AmiProvider.kt
@@ -1,0 +1,8 @@
+package com.atlassian.performance.tools.aws.api.ami
+
+import com.atlassian.performance.tools.aws.api.Aws
+
+interface AmiProvider {
+
+    fun provideAmiId(aws: Aws): String
+}