+++ CHANGELOG.md
@@ -23,6 +23,10 @@ Dropping a requirement of a major version of a dependency is a new contract.
 ## [Unreleased]
 [Unreleased]: https://bitbucket.org/atlassian/aws-resources/branches/compare/master%0Drelease-1.9.1
 
+### Added
+- Inject `AmiProvider` into `Aws` to choose the default AMI.
+- Expose `CanonicalAmiProvider` to find Ubuntu AMIs by name.
+
 ## [1.9.1] - 2023-01-24
 [1.9.1]: https://bitbucket.org/atlassian/aws-resources/branches/compare/release-1.9.1%0Drelease-1.9.0